pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestro-cli"
version = "0.2.1"
description = "Orchestro CLI — Playwright for Terminal Applications. Orchestrated CLI/TUI testing with screenshot-driven validation and intelligence-assisted scenario generation for Textual, Rich, and curses apps."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "VyB Project", email = "vyb@example.com"}
]
keywords = ["cli", "testing", "automation", "tui", "screenshot", "pexpect", "api", "graphql", "fastapi", "textual", "terminal", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pexpect>=4.8.0,<5.0.0",
    "pyyaml>=6.0,<7.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "strawberry-graphql[fastapi]>=0.200.0,<1.0.0",
    "httpx>=0.25.0,<1.0.0",
    "websockets>=12.0,<13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "mypy>=1.0",
    "httpx>=0.25.0",
]

[project.scripts]
orchestro = "orchestro_cli.cli:main"

[project.urls]
Homepage = "https://github.com/vyb/orchestro-cli"
Documentation = "https://github.com/vyb/orchestro-cli/blob/main/README.md"
Repository = "https://github.com/vyb/orchestro-cli"
Issues = "https://github.com/vyb/orchestro-cli/issues"

[tool.setuptools]
packages = ["orchestro_cli"]

[tool.setuptools.package-data]
orchestro_cli = ["py.typed"]
